[YITEAM-155] User, UserInfo 분리
