+++ src/main/java/com/yapp/ios1/controller/UserController.java
@@ -5,8 +5,6 @@ import com.yapp.ios1.model.user.User;
 import com.yapp.ios1.controller.dto.user.login.PasswordDto;
 import com.yapp.ios1.controller.dto.user.login.SignInDto;
 import com.yapp.ios1.controller.dto.user.login.SignUpDto;
-import com.yapp.ios1.model.user.Friend;
-import com.yapp.ios1.service.FriendService;
 import com.yapp.ios1.service.JwtService;
 import com.yapp.ios1.service.UserService;
 import com.yapp.ios1.annotation.Auth;
@@ -21,8 +19,6 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
-
-import java.util.List;
 
 import static com.yapp.ios1.message.ResponseMessage.*;
 
@@ -39,7 +35,6 @@ public class UserController {
     private final UserService userService;
     private final UserValidator userValidator;
     private final JwtService jwtService;
-    private final FriendService followService;
 
     @ApiOperation(value = "이메일 존재 여부")
     @GetMapping("/email-check")
@@ -75,30 +70,6 @@ public class UserController {
         Long userId = UserContext.getCurrentUserId();
         userService.changePassword(userId, passwordDto.getPassword());
         return ResponseEntity.ok(ResponseDto.of(HttpStatus.OK, CHANGE_PASSWORD_SUCCESS));
-    }
-
-    @ApiOperation(value = "마이 페이지")
-    @Auth
-    @GetMapping("/me")
-    public ResponseEntity<ResponseDto> getMyInfo() {
-        Long userId = UserContext.getCurrentUserId();
-        return ResponseEntity.ok(ResponseDto.of(HttpStatus.OK, GET_MY_INFO, userService.getUserInfo(userId)));
-    }
-
-    @ApiOperation(value = "사용자 페이지")
-    @Auth
-    @GetMapping("/{userId}")
-    public ResponseEntity<ResponseDto> getUserInfo(@PathVariable Long userId) {
-        Long currentUserId = UserContext.getCurrentUserId();
-        return ResponseEntity.ok(ResponseDto.of(HttpStatus.OK, GET_USER_INFO, userService.getOtherUserInfo(currentUserId, userId)));
-    }
-
-    // TODO User 가 지금 너무 거대해지고 Friend 모델을 사용하고 있기 때문에 FriendController로 이동하는 게 좋을 거 같음
-    @ApiOperation(value = "친구 리스트")
-    @GetMapping("/{userId}/friends")
-    public ResponseEntity<ResponseDto> getFriendList(@PathVariable Long userId) {
-        List<Friend> friendList = followService.getFriendList(userId);
-        return ResponseEntity.ok(ResponseDto.of(HttpStatus.OK, GET_FRIEND_LIST, friendList));
     }
 
     @ApiOperation(value = "회원 탈퇴")
